% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{evalue_change}
\alias{evalue_change}
\title{Evaluate Correction}
\usage{
evalue_change(otutab, set1, set2)
}
\arguments{
\item{otutab}{a character specifying the otutab file or a data.frame of
otutab. The first column should be ID. The first row should be header.}

\item{set1}{a character specifying the first dataset to compare.}

\item{set2}{a character specifying the second dataset to compare.}
}
\value{
a list contains p-value, mean, procrustes plot, bar plot
}
\description{
Using NMDS and procrustes analysis to evaluate the region effect
.
}
