% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fa_unique}
\alias{fa_unique}
\title{Unique fasta file}
\usage{
fa_unique(file, dir)
}
\arguments{
\item{file}{a character specifying the fasta file path}

\item{dir}{a character specifying the output dir}
}
\value{
No return value, only output files
}
\description{
Unique fasta file by label and sequence
}
