% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{model_call}
\alias{model_call}
\title{Call HMM model}
\usage{
model_call(gene = "18S", region = "v4", strand = "fwd")
}
\arguments{
\item{gene}{a character specifying the model target gene name, such as 18S,
5.8S, ITS.}

\item{region}{a character specifying the model target region of gene, such as
v1-9 for 18S, 5.8S for 5.8S, ITS1 and ITS2 for ITS.}

\item{strand}{a character specifying the model target strand of gene region,
such as forward(fwd) for all 18S regions, ITS1 and 5.8S, type 1-3 for ITS2.}
}
\value{
a character specifying the HMM model path
}
\description{
call hmm model in package
}
\examples{
model_call()
}
