% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer.R
\name{primer_search}
\alias{primer_search}
\title{Primer search}
\usage{
primer_search(keywords, target = "seq", ignore = NULL)
}
\arguments{
\item{keywords}{a vector of characters specifying the query words in any
order. But the letter case should be considered, such like using 18S not 18s,
using v4 not V4. The short words for direction are format strict, using fwd
for forward primer and rev for reverse primer.}

\item{target}{a vector of characters specifying the query result target
information. The character should use the builtin_primers list elements name,
such as gene, region, set, name, id, direction, specificity, seq,
start_yeast, reference. If the length of vector is 1, the result will be a
character. If the length of vector is more than 1, the result will be a list.}

\item{ignore}{a cector of characters specifying the ignored query range. The
character should use the builtin_primers list elements name, such as gene,
region, set, name, id, direction, specificity, seq, start_yeast, reference.}
}
\value{
a character or a vector of characters specifying the query target
information
}
\description{
Using usearch -search_pcr2 function to extract hypervariable
region based on primer alignment method.
}
\examples{
primer_v4_f = primer_search(keywords = c("18S","v4","fwd"))
}
