% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{range_scroller}
\alias{range_scroller}
\title{Gene region scroller}
\usage{
range_scroller(index)
}
\arguments{
\item{index}{a vector of numbers specifying the gene regions. 1-18S_v1,
2-18S_v2, 3-18S_v3, 4-18S_v4, 5-18S_v5, 6-18S_v6, 7-18S_v7, 8-18S_v8,
9-18S_v9, 10-ITS_ITS1, 11-5.8S_5.8S, 12-ITS_ITS2.}
}
\value{
a vector of characters specifying the pasted string of gene and
region name
}
\description{
Select gene region by a vector of index
}
\examples{
# single region
range_scroller(4)
# a range of regions
range_scroller(3:6)
}
