% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{region_extract_hmm}
\alias{region_extract_hmm}
\title{Extract Region by HMM}
\usage{
region_extract_hmm(seq, gene_region, extend = FALSE)
}
\arguments{
\item{seq}{a character specifying the template sequence file path.}

\item{gene_region}{a vector of characters specifying the gene name and region
name in a pasted string. This parameter usually defined by the range_scroller
function.}

\item{extend}{a logical specifying the extract method is or not extend to the
ending.}
}
\value{
No return value, only output files
}
\description{
Extract target region of gene by HMM model. The output is fasta
format sequences and txt format table.
}
