% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{run_usearch}
\alias{run_usearch}
\title{Usearch wrapper}
\usage{
run_usearch(command)
}
\arguments{
\item{command}{a character string specifying the usearch command, which start
with 'usearch'. In docker environment, single quote is defualt for command
transform. Double quote should be avoid to be used in command.}
}
\value{
stdout screen from usearch
}
\description{
Wrapper function for usearch. Check host local first, then check
docker container environment. Should setup the option parameter by the guide
in document.
}
\examples{
options(usearch.docker.path = "docker ps -aqf 'name=usearch' | \\\\
                               xargs -I {} docker exec -i {} /bin/sh -c ")
options(usearch.path = '/bin/')
run_usearch(command = "usearch -fastx_get_sample_names /mnt/Octocarol.fa \\\\
                               -output /mnt/Octocarol_samples.txt")
}
