% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer.R
\name{theor_length}
\alias{theor_length}
\title{Theoretical length of amplicon}
\usage{
theor_length(index)
}
\arguments{
\item{index}{a vector of numbers specifying the gene regions. 1-18S_v1,
2-18S_v2, 3-18S_v3, 4-18S_v4, 5-18S_v5, 6-18S_v6, 7-18S_v7, 8-18S_v8,
9-18S_v9, 10-ITS_ITS1, 11-5.8S_5.8S, 12-ITS_ITS2.}
}
\value{
a number specifying the theoretical length of amplicon
}
\description{
Query theoretical length of amplicon based on yeast location of
primer.
}
\examples{
theor_length(1:9)
}
