% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer.R
\name{usearch_pcr2}
\alias{usearch_pcr2}
\title{Usearch search_pcr2}
\usage{
usearch_pcr2(template, gene_region, fwd = "default", rev = "default")
}
\arguments{
\item{template}{a character specifying the template sequence file in fasta or
fastq format.}

\item{gene_region}{a vector of characters specifying the gene name and region
name in a pasted string. This parameter usually defined by the range_scroller
function.}

\item{fwd}{a character specifying the forward primer set name.}

\item{rev}{a character specifying the reverse primer set name.}
}
\value{
usearch output files in docker volumn
}
\description{
Using usearch -search_pcr2 function to extract gene region by
primer alignment method.
}
